[CDP-30] Compiler error cleanup

+++ cdp/cdp/src/engine/CpuEngine.cpp
@@ -28,12 +28,12 @@ CpuEngine::CpuEngine(std::vector<CUdeviceptr> depth_textures, std::vector<CUdevi
 
 	this->drone_alpha.Init();
 	this->drone_alpha.wallrider.height_texture = this->camera_list[0]->height_texture;
-	this->drone_alpha.rigidbody.position = Vector3{ 88, 0.4, 5 };
+	this->drone_alpha.rigidbody.position = Vector3{ 88.0f, 0.4f, 5.0f };
 	//this->drone_alpha.rigidbody.position = Vector3{ 24, 4, 34 };
 	//this->drone_alpha.rigidbody.position = Vector3{ 48, 4, 42 };
 
 	float rot_angle = -Math::Pi() / 4;
-	rot_angle = -(45.0 / 180.0) * Math::Pi();
+	rot_angle = -(45.0f / 180.0f) * Math::Pi();
 	Vector4 x_rot = Quaternion::QuaternionFromEulerParams(Vector3{ 0, 0, 1 }, rot_angle);
 	Vector4 quat = x_rot;// Quaternion::QuaternionFromEulerParams(Vector3{ 0, 1, 0 }, 3 * Math::Pi() / 4);
 	//quat = Quaternion::MultiplyQuaternions(quat, x_rot, false);
@@ -48,7 +48,7 @@ CpuEngine::CpuEngine(std::vector<CUdeviceptr> depth_textures, std::vector<CUdevi
 	Vector2 render_size{ this->world_space->space_data.world_size0.x, this->world_space->space_data.world_size0.z };
 	std::string save_path = "./data/results/path_confusion.jpg";
 
-	unsigned long long byte_count = render_size.x * render_size.y * 3;
+	unsigned long long byte_count = (unsigned long long)(render_size.x * render_size.y * 3);
 	this->simulation_image = new byte[byte_count];
 	memset(this->simulation_image, 0, byte_count);
 }
@@ -70,7 +70,7 @@ void CpuEngine::Update(GuiData* gui_data) {
 
 	std::chrono::steady_clock::duration frame_time_passed = std::chrono::steady_clock::now() - this->frame_begin_time;
 	unsigned long long time_lapsed = std::chrono::duration_cast<std::chrono::milliseconds>(frame_time_passed).count();
-	unsigned long long delta_time_steps = Physics::DeltaTimeMilli();
+	unsigned long long delta_time_steps = (unsigned long long)Physics::DeltaTimeMilli();
 
 	if (time_lapsed < delta_time_steps) {
 		unsigned long long sleep_time = delta_time_steps - time_lapsed;
@@ -146,7 +146,7 @@ void CpuEngine::PhysicsUpdate(GuiData* gui_data) {
 	this->drone_alpha.rigidbody.AirResistance(wind);
 	this->drone_alpha.rigidbody.Update();
 
-	this->drone_alpha.rigidbody.position = this->drone_alpha.rigidbody.position.Clip(Vector::ZERO3(), this->ihm_generator.world_size_strided - 0.1);
+	this->drone_alpha.rigidbody.position = this->drone_alpha.rigidbody.position.Clip(Vector::ZERO3(), this->ihm_generator.world_size_strided - 0.1f);
 }
 
 void CpuEngine::RenderUpdate(GuiData* gui_data) {
@@ -175,8 +175,8 @@ void CpuEngine::SaveSimulationImage(GuiData* gui_data) {
 	int k = 40;
 	for (int w = 0; w < render_size.x; w++) {
 		for (int h = 0; h < render_size.y; h++) {
-			Vector3 voxel_position{ w, k, h };
-			unsigned long long voxel_index = Indexer::FlatIndex3(w, k, h, this->world_space->space_data.world_size0.x, this->world_space->space_data.world_size0.y);
+			Vector3 voxel_position{ (float)w, (float)k, (float)h };
+			unsigned long long voxel_index = Indexer::FlatIndex3((float)w, (float)k, (float)h, this->world_space->space_data.world_size0.x, this->world_space->space_data.world_size0.y);
 			VoxelData voxel_data = world_voxels[voxel_index];
 
 			if (voxel_data.entity_id > 0) {
@@ -185,7 +185,7 @@ void CpuEngine::SaveSimulationImage(GuiData* gui_data) {
 		}
 	}
 
-	int result = stbi_write_jpg(save_path.c_str(), render_size.x, render_size.y, 3, this->simulation_image, 100);
+	int result = stbi_write_jpg(save_path.c_str(), (int)render_size.x, (int)render_size.y, 3, this->simulation_image, 100);
 
 	printf("Saved simulation image at: %s\n", save_path.c_str());
 }
@@ -227,11 +227,11 @@ void CpuEngine::SaveProximityHash(GuiData* gui_data) {
 		printf("\n\nWarning: File did not open when saving Proximity Hash:\n    %s!\n", save_path.c_str());
 		exit(1);
 	}
-	int result = fwrite(proximity_hash, sizeof(float), 16, out_file);
+	int result = (int)fwrite(proximity_hash, sizeof(float), (unsigned int)16, out_file);
 	fclose(out_file);
 
-	result = stbi_write_jpg((save_path + ".camera.jpg").c_str(), camera.phash_data_size.x, camera.phash_data_size.y, 1, forward_phash, 100);
-	result = stbi_write_jpg((save_path + ".height.jpg").c_str(), camera.phash_texture_size.x, camera.phash_texture_size.x, 4, height_texture, 100);
+	result = stbi_write_jpg((save_path + ".camera.jpg").c_str(), (int)camera.phash_data_size.x, (int)camera.phash_data_size.y, 1, forward_phash, 100);
+	result = stbi_write_jpg((save_path + ".height.jpg").c_str(), (int)camera.phash_texture_size.x, (int)camera.phash_texture_size.x, 4, height_texture, 100);
 
 	printf("    Done!\n\n");
 }
