+++ cdp/cdp/src/engine/cuda/CudaWorld.cu
@@ -31,9 +31,9 @@ VoxelData CudaWorld::ReadVoxel(SpaceData space_data, Vector3 position) {
 void CudaWorld::FillBox(SpaceBuilder space_builder, SpaceData space_data, VoxelData voxel_data, Vector3 lower, Vector3 width) {
     dim3 threads_per_block(4, 4, 4);
 
-    int x_blocks = ceil(width.x / (float)threads_per_block.x);
-    int y_blocks = ceil(width.y / (float)threads_per_block.y);
-    int z_blocks = ceil(width.z / (float)threads_per_block.z);
+    int x_blocks = (int)ceil(width.x / (float)threads_per_block.x);
+    int y_blocks = (int)ceil(width.y / (float)threads_per_block.y);
+    int z_blocks = (int)ceil(width.z / (float)threads_per_block.z);
 
     dim3 blocks_per_grid(x_blocks, y_blocks, z_blocks);
 
@@ -56,7 +56,7 @@ void CudaWorld::AssignAll(SpaceBuilder space_builder, SpaceData space_data, Voxe
 void CudaWorld::WritePoints(SpaceBuilder space_builder, SpaceData space_data, Vector3* points, unsigned long long point_count, VoxelData voxel_data) {
     dim3 threads_per_block(32, 1, 1);
 
-    int x_blocks = ceil(point_count / (float)threads_per_block.x);
+    int x_blocks = (int)ceil(point_count / (float)threads_per_block.x);
 
     dim3 blocks_per_grid(x_blocks, 1, 1);
 
@@ -69,7 +69,7 @@ void CudaWorld::WritePoints(SpaceBuilder space_builder, SpaceData space_data, Ve
 void CudaWorld::PlanarDensify(SpaceBuilder space_builder, SpaceData space_data, Vector3* points, unsigned long long point_count) {
     dim3 threads_per_block(512, 1, 1);
 
-    unsigned long long x_blocks = ceil(point_count / threads_per_block.x);
+    unsigned long long x_blocks = (unsigned long long)ceil(point_count / threads_per_block.x);
 
     dim3 blocks_per_grid(x_blocks, 1, 1);
 
@@ -83,13 +83,13 @@ void CudaWorld::StochasticSubtraction(SpaceBuilder space_builder, SpaceData spac
     std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
     dim3 threads_per_block(32, 1, 1);
 
-    int x_blocks = ceil(space_data.voxel_count0 / threads_per_block.x);
+    int x_blocks = (int)ceil(space_data.voxel_count0 / threads_per_block.x);
 
     dim3 blocks_per_grid(x_blocks, 1, 1);
 
     printf("    Stochastic Subtraction:\n");
-    printf("        Block Count : (%lld, %lld, %lld)\n", x_blocks, 1, 1);
-    printf("        Thread Count: (%lld, %lld, %lld)\n", threads_per_block.x, 1, 1);
+    printf("        Block Count : (%d, %d, %d)\n", x_blocks, 1, 1);
+    printf("        Thread Count: (%d, %d, %d)\n", threads_per_block.x, 1, 1);
     printf("        Progress (Max 20 *): ");
     CudaWorld::StochasticSubtraction_Kernel<<<blocks_per_grid, threads_per_block>>>(space_builder, space_data);
 
@@ -97,7 +97,7 @@ void CudaWorld::StochasticSubtraction(SpaceBuilder space_builder, SpaceData spac
     CudaError::CheckError((cudaError_enum)cudaDeviceSynchronize(), __FILE__, __LINE__);
 
     std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();
-    int time_lapsed = std::chrono::duration_cast<std::chrono::seconds>(end - begin).count();
+    int time_lapsed = (int)std::chrono::duration_cast<std::chrono::seconds>(end - begin).count();
     printf("\n        Done!\n");
     printf("            Time Elapsed: %d s\n", time_lapsed);
 }
