+++ cdp/cdp/src/entity/DroneAlpha.h
@@ -12,10 +12,10 @@ struct DroneAlpha {
 	float yaw_max = 0.5;
 
 	Vector3 drift_direction{ 0, -1, 0 };
-	float drift_size = 0.5;
-	float drift_delta = 0.4;
+	float drift_size = 0.5f;
+	float drift_delta = 0.4f;
 
-	unsigned long long seed = 555586;
+	long seed = 555586;
 
 	void Init() {
 		this->rigidbody.Init();
@@ -33,14 +33,14 @@ struct DroneAlpha {
 
 	void FollowCommand(Vector3 command) {
 		float yaw_target = -command.x * this->yaw_max;
-		this->rigidbody.angular_velocity.y = (0.4 * yaw_target) + (0.6 * this->rigidbody.angular_velocity.y);
+		this->rigidbody.angular_velocity.y = (0.4f * yaw_target) + (0.6f * this->rigidbody.angular_velocity.y);
 
 		Vector3 body_forward = this->rigidbody.Forward();
 		body_forward.y = 0;
 		Vector3 velocity_target = body_forward * this->forward_max * command.z;
 		
 		velocity_target.y += command.y * this->climb_max;
-		this->rigidbody.velocity = (velocity_target * 0.4) + (this->rigidbody.velocity * 0.6);
+		this->rigidbody.velocity = (velocity_target * 0.4f) + (this->rigidbody.velocity * 0.6f);
 
 		//Vector3 xz_velocity = body_forward * command.z * 2;
 
@@ -50,15 +50,15 @@ struct DroneAlpha {
 
 	void Drift() {
 		long noisy_bits = this->NextSample(this->seed);
-		float noise_x = (float)noisy_bits / 32768.0;
+		float noise_x = (float)noisy_bits / 32768.0f;
 		noise_x = (2 * noise_x) - 1;
 
 		noisy_bits = this->NextSample(noisy_bits);
-		float noise_y = (float)noisy_bits / 32768.0;
+		float noise_y = (float)noisy_bits / 32768.0f;
 		noise_y = (2 * noise_x) - 1;
 
 		noisy_bits = this->NextSample(noisy_bits);
-		float noise_z = (float)noisy_bits / 32768.0;
+		float noise_z = (float)noisy_bits / 32768.0f;
 		noise_z = (2 * noise_x) - 1;
 
 		Vector3 drift_offset{ noise_x, noise_y, noise_z };
